Change VoidPtr operations to use byte* pointer arithmetic; fix RSTM encoding and playback to work in 64-bit builds
With these changes, BRSTM encoding and playback should work in both 32-bit and 64-bit applications. (I used some of this for LoopingAudioConverter, because I wanted it to work in Mono on 64-bit.) Several other BrawlBox features still assume a 32-bit pointer size, so those would need to be changed if we wanted or needed to support a 64-bit build.